requirements.txt
streamlit
openai
langchain
langchain-community
faiss-cpu
PyMuPDF
PyPDF2
pdfplumber
fpdf
plotly>=6.1.1
pandas
matplotlib
yfinance
requests
Pillow
python-dotenv
reportlab
kaleido>=1.0.0

